Move clientID check out of Accept()
Gorpc server use clientID (which by default client IP), to dispatch client messages.
In case of Tyk it was causing issues, since when hybrid traffic goes though the proxy, behind MDCB, original IP gets lost, and all clients looks the same. It was causing issues when clients did not received messages in some cases. We solved it by introducing additional protocol handshake, on which Gateway sends its UUID, which will be used as clientID instead of IP.

We had to implement handshake functionality inside Accept() function, because of Gorpc design. Accept() in networking applications should be non blocking, but in our case we did a few connection reads inside it, which in most cases was fine (since they are fast), but in some cases it was blocking for a long period of time, because of network timeout and etc (for example in case of TCP health-checks).

This change moves clientID logic out of Accept() function to OnConnect() hook, and MDCB does its handshake logic inside it.

Additionally all connection variables set to type net.Conn.